requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
redis[hiredis]==5.0.0
asyncpg==0.29.0
prometheus-client==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
sqlalchemy[asyncio]==2.0.29
pydantic-settings==2.2.1
aiosqlite==0.20.0
